Add :: delete, edit post validation

+++ src/repositories/post.repository.js
@@ -3,6 +3,7 @@ const {getPayloadFromHeader} = require("../utils/jwt.util");
 const PostModel = require('../models/post.model')
 const UserModel = require('../models/user.model')
 const {getCurrentTime} = require("../utils/time");
+const {ApiExceptions} = require("../utils/exceptions/api.exceptions");
 
 class PostRepository {
     addPost = async (body, header) => {
@@ -28,13 +29,26 @@ class PostRepository {
         return await PostModel.findByUserId(user_id);
     }
 
-    deleteById = async (post_id) => {
+    deleteById = async (post_id, header) => {
+        const payload = await getPayloadFromHeader(header);
+        const {email} = payload;
+        const user = await UserModel.findByEmail(email);
+        const post = await PostModel.findById(post_id);
+        if (user.user_id !== post.user_id) {
+            throw new ApiExceptions('this post is not yours', 400);
+        }
         await PostModel.deleteById(post_id);
         return 'delete success';
     }
 
-    editPost = async ( body) => {
+    editPost = async (body, header) => {
         const {post_id, content, latitude, longitude, is_public, user_id} = body;
+        const payload = await getPayloadFromHeader(header);
+        const {email} = payload;
+        const user = await UserModel.findByEmail(email);
+        if (user.user_id !== user_id) {
+            throw new ApiExceptions('this post is not yours', 400);
+        }
         return await PostModel.update(post_id, content, latitude, longitude, is_public, user_id);
     }
 }
